refactor(backend): move app bootstrap hooks into lib for coverable main path
