+++ src-tauri/src/lib.rs
@@ -2,14 +2,22 @@ use std::sync::{Mutex, OnceLock};
 
 type BuilderFactory = fn() -> tauri::Builder<tauri::Wry>;
 type Runner = fn(tauri::Builder<tauri::Wry>);
+type BackendRunner = fn();
 
 const DEFAULT_BUILDER_FACTORY: BuilderFactory = build_app;
 const DEFAULT_RUNNER: Runner = run_builder;
+const DEFAULT_BACKEND_RUNNER: BackendRunner = run;
 
 fn run_hooks() -> &'static Mutex<(BuilderFactory, Runner)> {
     static RUN_HOOKS: OnceLock<Mutex<(BuilderFactory, Runner)>> = OnceLock::new();
 
     RUN_HOOKS.get_or_init(|| Mutex::new((DEFAULT_BUILDER_FACTORY, DEFAULT_RUNNER)))
+}
+
+fn backend_runner() -> &'static Mutex<BackendRunner> {
+    static BACKEND_RUNNER: OnceLock<Mutex<BackendRunner>> = OnceLock::new();
+
+    BACKEND_RUNNER.get_or_init(|| Mutex::new(DEFAULT_BACKEND_RUNNER))
 }
 
 pub fn greet(name: &str) -> String {
@@ -53,4 +61,26 @@ pub fn run() {
     let (builder_factory, runner) = *run_hooks().lock().expect("run hooks lock poisoned");
 
     run_with(builder_factory, runner);
+}
+
+pub fn run_application(run_backend: fn()) {
+    run_backend();
+}
+
+pub fn set_backend_runner(run_backend: BackendRunner) {
+    *backend_runner()
+        .lock()
+        .expect("backend runner lock poisoned") = run_backend;
+}
+
+pub fn reset_backend_runner() {
+    set_backend_runner(DEFAULT_BACKEND_RUNNER);
+}
+
+pub fn start_application() {
+    let run_backend = *backend_runner()
+        .lock()
+        .expect("backend runner lock poisoned");
+
+    run_application(run_backend);
 }
